Ejercicio3-7

+++ Ejercicio3.java
@@ -13,7 +13,7 @@ public class Ejercicio3 {
         }
     }
 
-    public static float factorial(int n) {
+    public static double factorial(int n) {
         if (n == 0) {
             return 1;
         } else {
